+++ configs/detection/vitdet/_base_/datasets/coco.py
@@ -15,7 +15,7 @@ CLASSES = [
 ]
 
 # dataset settings
-data_root = '/apsarapangu/disk2/yunji.cjy/coco/'
+data_root = 'data/coco/'
 img_norm_cfg = dict(
     mean=[123.675, 116.28, 103.53], std=[58.395, 57.12, 57.375], to_rgb=True)
