+++ easycv/core/optimizer/__init__.py
@@ -1,6 +1,7 @@
 from torch.optim import *
 
+from .adam import _AdamW
 from .builder import build_optimizer_constructor
+from .lars import LARS
 from .layer_decay_optimizer_constructor import LayerDecayOptimizerConstructor
-from .optimizer import LARS, _AdamW
 from .ranger import Ranger
