+++ easycv/utils/fp16_utils.py
@@ -1,3 +1,4 @@
+# Copyright (c) OpenMMLab. All rights reserved.
 from collections import abc
 
 import numpy as np
